150 made me so irritated.
It completes in *under a second* in my repl, why does it time out on
4Clojure? Does the sandbox hang on a certain operation or something? Does
Java interop explode it? Are the 4Clojure servers actually potatoes? Ah,
life is full of mysteries.

I just gave up and copied a solution I found on Stack Overflow into the
field because I feel perfectly justified in saying I've solved it, and
with a not particularly inefficient solution.

I mean, how does it go from under a second in my repl to like, more than
10 in 4Clojure?